Updated to 1.15 (@darkhekromant and @zigthehedge):
Fixed plug drop pos
Fixed all stack placement
Fixed crash with interact (check ItemAttributes)
Fixed pallet placement crash
Fixed PlugnFeatherBlock.OnBlockBroken crash
Fixed PlugnFeatherBlock.CheckDone crash
Fixed rubble hammer head recipe
Refactoring
Added ru lang
