+++ QuarryWorks/assets/quarryworks/blocktypes/rubblestorage/rubblestorage.json
@@ -10,17 +10,17 @@
     },
     {
       "code": "type",
-      "states": [ "empty", "stone", "gravel", "sand" ]
+      "states": ["empty", "stone", "gravel", "sand"]
     },
     {
       "code": "dir",
-      "states": [ "north", "south", "east", "west" ]
+      "states": ["north", "south", "east", "west"]
     }
   ],
 
   "creativeinventory": {
-    "general": [ "*-sandstone-empty-north" ],
-    "stoneworks": [ "*-sandstone-empty-north" ]
+    "general": ["*-sandstone-empty-north"],
+    "quarryworks": ["*-sandstone-empty-north"]
   },
   "attributesByType": {
     "*-east": {
@@ -35,7 +35,7 @@
           "x": 0,
           "y": 0,
           "z": 1,
-          "varType": [ [ "dir", "west" ] ]
+          "varType": [["dir", "west"]]
         }
       ]
     },
@@ -51,7 +51,7 @@
           "x": -1,
           "y": 0,
           "z": 0,
-          "varType": [ [ "dir", "north" ] ]
+          "varType": [["dir", "north"]]
         }
       ]
     },
@@ -67,7 +67,7 @@
           "x": 1,
           "y": 0,
           "z": 0,
-          "varType": [ [ "dir", "south" ] ]
+          "varType": [["dir", "south"]]
         }
       ]
     },
@@ -83,9 +83,7 @@
           "x": 0,
           "y": 0,
           "z": -1,
-          "varType": [
-            [ "dir", "east" ]
-          ]
+          "varType": [["dir", "east"]]
         }
       ]
     }
@@ -307,7 +305,6 @@
     "all": true
   },
   "lightabsorption": 0,
-
 
   "heldTpIdleAnimation": "holdbothhandslarge",
   "heldTpUseAnimation": "twohandplaceblock",
