+++ QuarryWorks/assets/quarryworks/blocktypes/stoneStorage/med.json
@@ -10,12 +10,11 @@
     },
     {
       "code": "dir",
-      "states": [ "north", "south", "east", "west" ]
+      "states": ["north", "south", "east", "west"]
     }
   ],
   "creativeinventory": {
-    "general": [ "*-north" ],
-    "stoneworks": [ "*-north" ]
+    "general": ["*-north"]
   },
   "attributesByType": {
     "*-east": {
@@ -24,7 +23,7 @@
           "x": 0,
           "y": 0,
           "z": 1,
-          "varType": [ [ "dir", "west" ] ]
+          "varType": [["dir", "west"]]
         }
       ]
     },
@@ -34,7 +33,7 @@
           "x": -1,
           "y": 0,
           "z": 0,
-          "varType": [ [ "dir", "north" ] ]
+          "varType": [["dir", "north"]]
         }
       ]
     },
@@ -44,7 +43,7 @@
           "x": 1,
           "y": 0,
           "z": 0,
-          "varType": [ [ "dir", "south" ] ]
+          "varType": [["dir", "south"]]
         }
       ]
     },
@@ -54,9 +53,7 @@
           "x": 0,
           "y": 0,
           "z": -1,
-          "varType": [
-            [ "dir", "east" ]
-          ]
+          "varType": [["dir", "east"]]
         }
       ]
     }
@@ -96,7 +93,6 @@
     "all": true
   },
   "lightabsorption": 0,
-
 
   "heldTpIdleAnimation": "holdbothhandslarge",
   "heldTpUseAnimation": "twohandplaceblock",
