+++ QuarryWorks/assets/quarryworks/itemtypes/debugger/slabstacksetter.json
@@ -6,15 +6,14 @@
     "slabstackdb": true
   },
 
-
   "shape": { "base": "items/tools/ballodirt" },
   "textures": {
-    "all": { "base": "game:block/metal/plate/copper" },
+    "all": { "base": "game:block/metal/plate/copper" }
   },
 
   "creativeinventory": {
-    "general": [ "*" ],
-    "stoneworks": [ "*" ]
+    "general": ["*"],
+    "quarryworks": ["*"]
   },
 
   "heldTpUseAnimation": "axechop",
