+++ QuarryWorks/assets/quarryworks/itemtypes/finishingchisel.json
@@ -3,7 +3,17 @@
   "variantgroups": [
     {
       "code": "material",
-      "states": [ "copper", "tinbronze", "bismuthbronze", "blackbronze", "gold", "silver", "iron", "meteoriciron", "steel" ]
+      "states": [
+        "copper",
+        "tinbronze",
+        "bismuthbronze",
+        "blackbronze",
+        "gold",
+        "silver",
+        "iron",
+        "meteoriciron",
+        "steel"
+      ]
     }
   ],
   "attributes": {
@@ -43,8 +53,8 @@
   },
 
   "creativeinventory": {
-    "general": [ "*" ],
-    "stoneworks": [ "*" ]
+    "general": ["*"],
+    "quarryworks": ["*"]
   },
 
   "heldTpUseAnimation": "knap",
