+++ QuarryWorks/assets/quarryworks/recipes/smithing/rubblehammerhead.json
@@ -8,14 +8,14 @@
   "pattern": [
     [
       "######",
-      "_#####",
-      "_#####",
+      "__####",
+      "__####",
       "######"
     ],
     [
       "######",
-      "_#####",
-      "_#####",
+      "__####",
+      "__####",
       "######"
     ]
   ],
