+++ QuarryWorks/modinfo.json
@@ -1,11 +1,11 @@
 {
   "type": "code",
-  "modid": "stoneworks",
-  "name": "quarry works",
-  "author": "Wraith",
-  "description": "Testing!",
-  "version": "1.2.4",
-  "dependency": {
-    "game": ""
+  "modid": "quarryworks",
+  "name": "Quarry Works!",
+  "authors": ["Wraith"],
+  "description": "",
+  "version": "2.0.7",
+  "dependencies": {
+    "game": "1.15.0"
   }
 }
